fix: the Sonar findings, and one of them was real
The quality gate was failing on reliability, and the reason was a single bug:
Gemini's usage added two int token counts and widened the result, so a long
addition was done in int (S2184). Cast, and the gate's new_reliability_rating
goes back to 1.

The rest are smells, fixed where they were right:

- What a summary attempt came to is said once. "written", "nothing",
  "lease-refused", "fault" and "empty" were literals repeated across both
  summarisers; they are constants on SummaryObservation now, next to the tag
  they are reported under, because a dashboard groups by them and two copies of
  a vocabulary is how the two drift apart.
- SummaryObservation's `kind` field no longer sits beside a `KIND` constant
  meaning something else: the tag keys are KIND_TAG and OUTCOME_TAG.
- TokenUsageHandler's private `record` is `sample`, since `record` is a
  restricted identifier.
- "usage must not be null" is one constant on InferenceResult, said by all four
  arms.
- An unnamed pattern needs no type: `var _` is `_` in the record patterns that
  ignore a component.
- Tests: a lambda that was a method reference, and an anonymous
  ObservationHandler that was a lambda.

+++ nessy-inference/gemini/src/main/java/org/jwcarman/nessy/inference/gemini/GeminiInferenceProvider.java
@@ -148,7 +148,7 @@ public final class GeminiInferenceProvider implements InferenceProvider, AutoClo
             counted ->
                 new Usage(
                     counted.promptTokenCount().orElse(0),
-                    counted.candidatesTokenCount().orElse(0)
+                    (long) counted.candidatesTokenCount().orElse(0)
                         + counted.thoughtsTokenCount().orElse(0)))
         .orElse(Usage.unknown());
   }
